feat: release v0.19.11-no-telemetry — align with upstream main, resolve merge conflicts, verify no-telemetry policy
Merge upstream main (v0.19.11) into dev branch. Resolved 110+ merge
conflicts across package.json files, telemetry modules, CLI serve routes,
web-shell components, SDK types, and documentation.

No-telemetry policy enforcement:
- Removed all @opentelemetry/* dependencies from packages/core/package.json
- loggers.ts: all log functions are no-ops except logApiResponse, logApiError,
  logToolCall which forward to local-only uiTelemetryService (EventEmitter)
- daemon-tracing.ts: imports from ./dummy-otel.js (not @opentelemetry/api)
- qwen-logger/qwen-logger.ts: all methods are no-op stubs
- session-tracing.ts: no-op stub with any types
- Installation ID hardcoded to 00000000-0000-0000-0000-000000000000
- enableAutoUpdate hardcoded to false
- gitCoAuthor hardcoded to { commit: false, pr: false }
- Version references updated: Dockerfile, install.sh, README.md -> v0.19.11

Test fixes:
- server.test.ts: skipped OTel span test (requires real OTel context)
- capabilities-docs-contract.test.ts: took upstream docs with markers
- run-qwen-serve.test.ts: fixed async mock for initializeTelemetry

Build: clean, 0 errors
Core tests: 15115 passed
CLI tests: 6095 passed (4 pre-existing env failures, 2 fixed)

+++ packages/sdk-typescript/src/types/queryOptionsSchema.ts
@@ -217,7 +217,7 @@ export const QueryOptionsSchema = z
       >((val) => typeof val === 'function', { message: 'stderr must be a function' })
       .optional(),
     logLevel: z.enum(['debug', 'info', 'warn', 'error']).optional(),
-    maxSessionTurns: z.number().optional(),
+    maxSessionTurns: z.number().int().optional(),
     coreTools: z.array(z.string()).optional(),
     excludeTools: z.array(z.string()).optional(),
     allowedTools: z.array(z.string()).optional(),
